initial bg and border rendering

+++ src/utils/mod.rs
@@ -40,9 +40,9 @@ pub fn open_file(filename: &str, offset: u64) -> Vec<u8>
 
 // set 8 consecutive buffer elements to single value for faster update of
 // a single 8-pixel screen chunk
-pub fn memset8(buffer: &mut [u32], start: usize, value: u8)
+pub fn memset8(buffer: &mut [u32], start: usize, value: u32)
 {
-    buffer[start]   = value as u32;
+    buffer[start]   = value;
     buffer[start+1] = buffer[start];
     buffer[start+2] = buffer[start];
     buffer[start+3] = buffer[start];
